+++ src/Zicht/Itertools/twig/Extension.php
@@ -5,44 +5,59 @@
 
 namespace Zicht\Itertools\twig;
 
+use Twig\Extension\GlobalsInterface;
 use Zicht\Itertools;
 
 /**
  * Twig extension.
  */
-class Extension extends \Twig_Extension
+class Extension extends \Twig_Extension implements GlobalsInterface
 {
+    /**
+     * {@inheritDoc}
+     */
+    public function getGlobals()
+    {
+        return [
+            'itf' => new Itertools\util\Filters(),
+            'itm' => new Itertools\util\Mappings(),
+            'itr' => new Itertools\util\Reductions(),
+        ];
+    }
+
     /**
      * {@inheritDoc}
      */
     public function getFilters()
     {
         return [
-            // filter names are case-sensitive
-            new \Twig_SimpleFilter('all', '\Zicht\Itertools\all'),
-            new \Twig_SimpleFilter('any', '\Zicht\Itertools\any'),
-            new \Twig_SimpleFilter('chain', '\Zicht\Itertools\chain'),
-            new \Twig_SimpleFilter('collapse', '\Zicht\Itertools\collapse'),
-            new \Twig_SimpleFilter('filter', [$this, 'filter']),
-            new \Twig_SimpleFilter('first', '\Zicht\Itertools\first'),
-            new \Twig_SimpleFilter('group_by', [$this, 'groupBy']),
-            new \Twig_SimpleFilter('last', '\Zicht\Itertools\last'),
-            new \Twig_SimpleFilter('map', [$this, 'map']),
-            new \Twig_SimpleFilter('map_by', [$this, 'mapBy']),
-            new \Twig_SimpleFilter('reduce', '\Zicht\Itertools\reduce'),
-            new \Twig_SimpleFilter('reversed', '\Zicht\Itertools\reversed'),
-            new \Twig_SimpleFilter('sorted', [$this, 'sorted']),
-            new \Twig_SimpleFilter('unique', [$this, 'unique']),
-            new \Twig_SimpleFilter('zip', '\Zicht\Itertools\zip'),
+            new \Twig_SimpleFilter('it', [$this, 'it']),
+
+            // deprecated filters (because 'it' was introduced)
+            new \Twig_SimpleFilter('all', '\Zicht\Itertools\all', ['deprecated' => true, 'alternative' => '|it.all']),
+            new \Twig_SimpleFilter('any', '\Zicht\Itertools\any', ['deprecated' => true, 'alternative' => '|it.any']),
+            new \Twig_SimpleFilter('chain', '\Zicht\Itertools\chain', ['deprecated' => true, 'alternative' => '|it.chain']),
+            new \Twig_SimpleFilter('collapse', '\Zicht\Itertools\collapse', ['deprecated' => true, 'alternative' => '|it.collapse']),
+            new \Twig_SimpleFilter('filter', [$this, 'filter'], ['deprecated' => true, 'alternative' => '|it.filter']),
+            new \Twig_SimpleFilter('first', '\Zicht\Itertools\first', ['deprecated' => true, 'alternative' => '|it.first']),
+            new \Twig_SimpleFilter('group_by', [$this, 'groupBy'], ['deprecated' => true, 'alternative' => '|it.groupBy']),
+            new \Twig_SimpleFilter('last', '\Zicht\Itertools\last', ['deprecated' => true, 'alternative' => '|it.last']),
+            new \Twig_SimpleFilter('map', [$this, 'map'], ['deprecated' => true, 'alternative' => '|it.map']),
+            new \Twig_SimpleFilter('map_by', [$this, 'mapBy'], ['deprecated' => true, 'alternative' => '|it.mapBy']),
+            new \Twig_SimpleFilter('reduce', '\Zicht\Itertools\reduce', ['deprecated' => true, 'alternative' => '|it.reduce']),
+            new \Twig_SimpleFilter('reversed', '\Zicht\Itertools\reversed', ['deprecated' => true, 'alternative' => '|it.reversed']),
+            new \Twig_SimpleFilter('sorted', [$this, 'sorted'], ['deprecated' => true, 'alternative' => '|it.sorted']),
+            new \Twig_SimpleFilter('unique', [$this, 'unique'], ['deprecated' => true, 'alternative' => '|it.unique']),
+            new \Twig_SimpleFilter('zip', '\Zicht\Itertools\zip', ['deprecated' => true, 'alternative' => '|it.zip']),
 
             // deprecated filters
-            new \Twig_SimpleFilter('filterby', [$this, 'deprecatedFilterBy'], ['deprecated' => true, 'alternative' => 'filter']),
-            new \Twig_SimpleFilter('groupBy', [$this, 'deprecatedGroupBy'], ['deprecated' => true, 'alternative' => 'group_by']),
-            new \Twig_SimpleFilter('groupby', [$this, 'deprecatedGroupBy'], ['deprecated' => true, 'alternative' => 'group_by']),
-            new \Twig_SimpleFilter('mapBy', [$this, 'deprecatedMapBy'], ['deprecated' => true, 'alternative' => 'map_by']),
-            new \Twig_SimpleFilter('mapby', [$this, 'deprecatedMapBy'], ['deprecated' => true, 'alternative' => 'map_by']),
-            new \Twig_SimpleFilter('sum', [$this, 'deprecatedSum'], ['deprecated' => true, 'alternative' => 'reduce']),
-            new \Twig_SimpleFilter('uniqueby', [$this, 'deprecatedUniqueBy'], ['deprecated' => true, 'alternative' => 'unique']),
+            new \Twig_SimpleFilter('filterby', [$this, 'deprecatedFilterBy'], ['deprecated' => true, 'alternative' => '|it.filter']),
+            new \Twig_SimpleFilter('groupBy', [$this, 'deprecatedGroupBy'], ['deprecated' => true, 'alternative' => '|it.groupBy']),
+            new \Twig_SimpleFilter('groupby', [$this, 'deprecatedGroupBy'], ['deprecated' => true, 'alternative' => '|it.groupBy']),
+            new \Twig_SimpleFilter('mapBy', [$this, 'deprecatedMapBy'], ['deprecated' => true, 'alternative' => '|it.mapBy']),
+            new \Twig_SimpleFilter('mapby', [$this, 'deprecatedMapBy'], ['deprecated' => true, 'alternative' => '|it.mapBy']),
+            new \Twig_SimpleFilter('sum', [$this, 'deprecatedSum'], ['deprecated' => true, 'alternative' => '|it.reduce']),
+            new \Twig_SimpleFilter('uniqueby', [$this, 'deprecatedUniqueBy'], ['deprecated' => true, 'alternative' => '|it.unique']),
         ];
     }
 
@@ -52,18 +67,30 @@ class Extension extends \Twig_Extension
     public function getFunctions()
     {
         return [
-            new \Twig_SimpleFunction('chain', '\Zicht\Itertools\chain'),
-            new \Twig_SimpleFunction('first', '\Zicht\Itertools\first'),
-            new \Twig_SimpleFunction('last', '\Zicht\Itertools\last'),
+            // deprecated functions (because 'it' was introduced)
+            new \Twig_SimpleFunction('chain', '\Zicht\Itertools\chain', ['deprecated' => true, 'alternative' => '|it.chain']),
+            new \Twig_SimpleFunction('first', '\Zicht\Itertools\first', ['deprecated' => true, 'alternative' => '|it.first']),
+            new \Twig_SimpleFunction('last', '\Zicht\Itertools\last', ['deprecated' => true, 'alternative' => '|it.last']),
 
-            // functions to create closures
-            new \Twig_SimpleFunction('reducing', [$this, 'reducing']),
-            new \Twig_SimpleFunction('mapping', [$this, 'mapping']),
-            new \Twig_SimpleFunction('filtering', [$this, 'filtering']),
+            // deprecated functions (because 'itr', 'itm', and 'itf' were introduced)
+            new \Twig_SimpleFunction('reducing', [$this, 'reducing'], ['deprecated' => true, 'alternative' => 'itr']),
+            new \Twig_SimpleFunction('mapping', [$this, 'mapping'], ['deprecated' => true, 'alternative' => 'itm']),
+            new \Twig_SimpleFunction('filtering', [$this, 'filtering'], ['deprecated' => true, 'alternative' => 'itf']),
 
             // deprecated functions
             new \Twig_SimpleFunction('reduction', [$this, 'deprecatedGetReduction'], ['deprecated' => true, 'alternative' => 'reducing']),
         ];
+    }
+
+    /**
+     * Takes an iterable and returns an object that allow mapping, sorting, etc.
+     *
+     * @param $iterable
+     * @return Itertools\lib\Interfaces\FiniteIterableInterface|Itertools\lib\IterableIterator
+     */
+    public function it($iterable)
+    {
+        return Itertools\iterable($iterable);
     }
 
     /**
@@ -72,6 +99,7 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param mixed $strategy
      * @return Itertools\lib\UniqueIterator
+     * @deprecated
      */
     public function unique($iterable, $strategy = null)
     {
@@ -92,6 +120,7 @@ class Extension extends \Twig_Extension
      * @param string|\Closure $closure
      * @param mixed $initializer
      * @return mixed
+     * @deprecated
      */
     public function reduce($iterable, $closure = 'add', $initializer = null)
     {
@@ -107,6 +136,7 @@ class Extension extends \Twig_Extension
      * @param string|\Closure $strategy
      * @param boolean $sort
      * @return Itertools\lib\GroupbyIterator
+     * @deprecated
      */
     public function groupBy($iterable, $strategy, $sort = true)
     {
@@ -120,6 +150,7 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param null $strategy
      * @return Itertools\lib\FilterIterator
+     * @deprecated
      */
     public function filter($iterable, $strategy = null)
     {
@@ -134,6 +165,7 @@ class Extension extends \Twig_Extension
      * @param string|\Closure $strategy
      * @param bool $reverse
      * @return Itertools\lib\SortedIterator
+     * @deprecated
      */
     public function sorted($iterable, $strategy = null, $reverse = false)
     {
@@ -146,6 +178,7 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param string|\Closure $strategy
      * @return Itertools\lib\MapIterator
+     * @deprecated
      */
     public function map($iterable, $strategy)
     {
@@ -159,6 +192,7 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param string|\Closure $strategy
      * @return Itertools\lib\MapByIterator
+     * @deprecated
      */
     public function mapBy($iterable, $strategy)
     {
@@ -181,24 +215,11 @@ class Extension extends \Twig_Extension
         // note, once we stop supporting php 5.5, we can rewrite the code below
         // to the reducing($name, ...$args) structure.
         // http://php.net/manual/en/functions.arguments.php#functions.variable-arg-list
-
-        if (is_string($name) &&
-            in_array(
-                $name,
-                [
-                    'add',
-                    'chain',
-                    'join',
-                    'max',
-                    'min',
-                    'mul',
-                    'sub',
-                ]
-            )) {
-            return call_user_func_array(sprintf('\Zicht\Itertools\reductions\%s', $name), array_slice(func_get_args(), 1));
+        $method = sprintf('\Zicht\Itertools\util\Reductions::%s', $name);
+        if (!is_callable($method)) {
+            throw new \InvalidArgumentException(sprintf('$name "%s" is not a valid reduction.', $name));
         }
-
-        throw new \InvalidArgumentException(sprintf('$NAME "%s" is not a valid reduction.', $name));
+        return call_user_func_array($method, array_slice(func_get_args(), 1));
     }
 
     /**
@@ -216,29 +237,11 @@ class Extension extends \Twig_Extension
         // note, once we stop supporting php 5.5, we can rewrite the code below
         // to the reducing($name, ...$args) structure.
         // http://php.net/manual/en/functions.arguments.php#functions.variable-arg-list
-
-        if (is_string($name) &&
-            in_array(
-                $name,
-                [
-                    'json_decode',
-                    'json_encode',
-                    'key',
-                    'length',
-                    'lower',
-                    'lstrip',
-                    'random',
-                    'rstrip',
-                    'select',
-                    'strip',
-                    'type',
-                    'upper',
-                ]
-            )) {
-            return call_user_func_array(sprintf('\Zicht\Itertools\mappings\%s', $name), array_slice(func_get_args(), 1));
+        $method = sprintf('\Zicht\Itertools\util\Mappings::%s', $name);
+        if (!is_callable($method)) {
+            throw new \InvalidArgumentException(sprintf('$name "%s" is not a valid mapping.', $name));
         }
-
-        throw new \InvalidArgumentException(sprintf('$NAME "%s" is not a valid mapping.', $name));
+        return call_user_func_array($method, array_slice(func_get_args(), 1));
     }
 
     /**
@@ -256,23 +259,11 @@ class Extension extends \Twig_Extension
         // note, once we stop supporting php 5.5, we can rewrite the code below
         // to the reducing($name, ...$args) structure.
         // http://php.net/manual/en/functions.arguments.php#functions.variable-arg-list
-
-        if (is_string($name) &&
-            in_array(
-                $name,
-                [
-                    'equals',
-                    'in',
-                    'match',
-                    'not',
-                    'not_in',
-                    'type',
-                ]
-            )) {
-            return call_user_func_array(sprintf('\Zicht\Itertools\filters\%s', $name), array_slice(func_get_args(), 1));
+        $method = sprintf('\Zicht\Itertools\util\Filters::%s', $name);
+        if (!is_callable($method)) {
+            throw new \InvalidArgumentException(sprintf('$name "%s" is not a valid filter.', $name));
         }
-
-        throw new \InvalidArgumentException(sprintf('$NAME "%s" is not a valid filter.', $name));
+        return call_user_func_array($method, array_slice(func_get_args(), 1));
     }
 
     /**
@@ -282,14 +273,12 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param string|\Closure $strategy
      * @return Itertools\lib\FilterIterator
-     *
      * @deprecated Use filter instead!
      */
     public function deprecatedFilterBy($iterable, $strategy)
     {
         return Itertools\filter($strategy, $iterable);
     }
-
 
     /**
      * Make an iterator that returns consecutive groups from the
@@ -299,7 +288,6 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param string|\Closure $strategy
      * @return Itertools\lib\GroupbyIterator
-     *
      * @deprecated Use group_by instead!
      */
     public function deprecatedGroupBy($iterable, $strategy)
@@ -314,7 +302,6 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param string|\Closure $strategy
      * @return Itertools\lib\MapByIterator
-     *
      * @deprecated Use map_by instead!
      */
     public function deprecatedMapBy($iterable, $strategy)
@@ -328,7 +315,6 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param int $default
      * @return int
-     *
      * @deprecated Use reduce instead!
      */
     public function deprecatedSum($iterable, $default = 0)
@@ -345,7 +331,6 @@ class Extension extends \Twig_Extension
      * @param array|string|\Iterator $iterable
      * @param mixed $strategy
      * @return Itertools\lib\UniqueIterator
-     *
      * @deprecated Use unique instead!
      */
     public function deprecatedUniqueBy($iterable, $strategy = null)
@@ -359,7 +344,6 @@ class Extension extends \Twig_Extension
      * @param string $name
      * @return \Closure
      * @throws \InvalidArgumentException
-     *
      * @deprecated Use reducing instead!
      */
     public function deprecatedGetReduction($name)